All logic for website finished
a. A homepage only with the list of posts. This page should have the following sections:
i. A menu with all the tags available. The user should be able to filter the post list by tag.
ii. A feature section with the last 5 posts.
iii. A regular section with the remaining posts.
iv. All posts ordered from last to first (DESC).
b. A create/edit page. Each post should have the following information:
i. Image (Add a new input in the form in order to save an image url for the post.)
ii. Title
iii. Body
iv. Tags (list of tags separated by commas)
c. A single page for each post.
d. The user can search by title.
e. The user can delete a post. (Display a link or icon for delete)
f. The user can edit a post. (Display a link or icon for edit)